+++ example/ionic-angular-v5/src/app/app.module.ts
@@ -4,13 +4,11 @@ import { RouteReuseStrategy } from '@angular/router';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
-import { createErrorHandler, init as sentryAngularInit } from '@sentry/angular';
+import { createErrorHandler, init as sentryAngularInit } from '@sentry/angular-ivy';
 import * as Sentry from '@sentry/capacitor';
-import { Integrations } from '@sentry/tracing';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { Replay } from "@sentry/replay";
 
 // ATTENTION: Change the DSN below with your own to see the events in Sentry. Get one at sentry.io
 Sentry.init(
@@ -25,8 +23,8 @@ Sentry.init(
     enabled: true,
     // Use the tracing integration to see traces and add performance monitoring
     integrations: [
-      new Integrations.BrowserTracing(),
-      new Replay({
+      new Sentry.BrowserTracing(),
+      new Sentry.Replay({
         maskAllText: false,
         blockAllMedia: true,
       }),