+++ example/ionic-angular/src/app/app.module.ts
@@ -1,16 +1,16 @@
-import { ErrorHandler, NgModule } from '@angular/core';
+import { APP_INITIALIZER, ErrorHandler, NgModule } from "@angular/core";
 import { BrowserModule } from '@angular/platform-browser';
 import { RouteReuseStrategy } from '@angular/router';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
 import { IonicModule, IonicRouteStrategy } from '@ionic/angular';
-import { createErrorHandler, init as sentryAngularInit } from '@sentry/angular';
+import { Router } from "@angular/router";
+
+import { createErrorHandler, TraceService, routingInstrumentation, init as sentryAngularInit } from '@sentry/angular-ivy';
 import * as Sentry from '@sentry/capacitor';
-import { Integrations } from '@sentry/tracing';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { Replay } from "@sentry/replay";
 
 // ATTENTION: Change the DSN below with your own to see the events in Sentry. Get one at sentry.io
 Sentry.init(
@@ -25,8 +25,8 @@ Sentry.init(
     enabled: true,
     // Use the tracing integration to see traces and add performance monitoring
     integrations: [
-      new Integrations.BrowserTracing(),
-      new Replay({
+      new Sentry.BrowserTracing({routingInstrumentation}),
+      new Sentry.Replay({
         maskAllText: false,
         blockAllMedia: true,
       }),
@@ -62,6 +62,17 @@ Sentry.init(
       provide: ErrorHandler,
       useValue: createErrorHandler(),
     },
+    {
+      provide: TraceService,
+      deps: [Router],
+    },
+    {
+      provide: APP_INITIALIZER,
+      useFactory: () => () => {},
+      deps: [TraceService],
+      multi: true,
+    },
+
   ],
   bootstrap: [AppComponent]
 })